requirements.txt
# requirements.txt
#
# 生产环境依赖 (用于Docker镜像)
# ----------------------------------------------------

# Web 服务框架
fastapi
uvicorn[standard]
python-multipart

# AI 模型推理核心库
# 注意：torch版本应与生产服务器的CUDA版本匹配
torch --extra-index-url https://download.pytorch.org/whl/cu121
transformers
openai-whisper
sentence-transformers
accelerate
bitsandbytes